+++ src/main/java/roomescape/auth/infrastructure/JwtTokenProvider.java
@@ -4,7 +4,7 @@ import io.jsonwebtoken.*;
 import io.jsonwebtoken.security.Keys;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Component;
-import roomescape.auth.entity.Member;
+import roomescape.entity.Member;
 
 import java.security.Key;
 import java.util.Date;
