+++ src/main/java/roomescape/entity/Role.java
@@ -1,4 +1,4 @@
-package roomescape.auth.entity;
+package roomescape.entity;
 
 public enum Role {
     USER, ADMIN
